requirements.txt
# Veri İşleme ve Analiz
pandas
numpy

# Makine Öğrenmesi
scikit-learn
joblib

# Görselleştirme
plotly
matplotlib
seaborn
altair

# Web Arayüzü
streamlit

# Yapılandırma ve Yardımcı Araçlar
python-dotenv
pyyaml